Multiple Front-end changes
Added tables for playlist view, import, export, and merge pages to be populated with playlists once connections are setup.  Added functionality to display the current user in the top left of the screen.  Added buttons for merging, importing, and exporting that will send msgs to the server to perform each functionality accordingly.  Also changed a lot of formatting issues.